// src/types/enum.ts
export enum EStatus {
    IDLE = "idle",
    LOADING = "loading",
    ERROR = "error",
}

export enum EProduct {
    productLoadedFailes = "product/productLoadedFailes",
    productLoadingStarted = "product/productLoadingStarted",
    productLoadedSuccess = "product/productLoadedSuccess",
    productFilterCategories = "product/productFilterCategories",
    productSearch = "product/productSearch",
}

export enum EShop {
    shopAddCard = "shop/shopAddCard",
    shopRemoveCard = "shop/shopRemoveCard",
}